Make sure there are no certs & keys present in /etc/shibboleth in our base image.
It seems that installing shibboleth generates some certs and keys
which we are going to require people to override in shibboleth2.xml
anyways. However, to make sure anyone does not accidentally use these
for anything serious, we remove them. Also some image scanners shout
at us about the private keys in image layers (understandably so)
and we don't like that noise.