No longer serving acs version of this.
ACS (Azure Container Service) versions are no longer thought of as scope
(if they ever were) of this repository.
a) It's not OpenShift, so differences will bite us sooner or later
b) No real visibility for us to test and verify the images when we need to
   rebuild them in order to fix security issues
c) And so on. You can still look into what it looked like in previous commit
   if you want to see how it was.